# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample API",
    "version": "1.0.0",
    "description": "사용자와 제품 관리를 위한 CRUD API"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "개발 서버"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "사용자 관리 API"
    },
    {
      "name": "Products",
      "description": "제품 관리 API"
    }
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "모든 사용자 목록 조회",
        "description": "등록된 모든 사용자의 목록을 반환합니다.",
        "responses": {
          "200": {
            "description": "성공적으로 사용자 목록을 반환함",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "새 사용자 생성",
        "description": "새로운 사용자를 생성합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "사용자 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "특정 사용자 조회",
        "description": "특정 ID를 가진 사용자의 정보를 반환합니다.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "조회할 사용자의 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 정보를 성공적으로 반환함",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "모든 제품 목록 조회",
        "description": "등록된 모든 제품의 목록을 반환합니다.",
        "responses": {
          "200": {
            "description": "성공적으로 제품 목록을 반환함",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "특정 제품 조회",
        "description": "특정 ID를 가진 제품의 정보를 반환합니다.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "조회할 제품의 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "제품 정보를 성공적으로 반환함",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "제품을 찾을 수 없음"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "자동 생성되는 사용자 ID"
          },
          "name": {
            "type": "string",
            "description": "사용자 이름"
          },
          "email": {
            "type": "string",
            "description": "사용자 이메일"
          }
        }
      },
      "Product": {
        "type": "object",
        "required": [
          "name",
          "price",
          "stock",
          "category"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "자동 생성되는 제품 ID"
          },
          "name": {
            "type": "string",
            "description": "제품 이름"
          },
          "price": {
            "type": "number",
            "description": "제품 가격"
          },
          "stock": {
            "type": "integer",
            "description": "재고 수량"
          },
          "category": {
            "type": "string",
            "description": "제품 카테고리"
          }
        }
      }
    }
  }
}